+++ dbx/templates/projects/python_basic/render/hooks/post_gen_project.py
@@ -93,7 +93,7 @@ class PostProcessor:
 
         configure(
             environment="default", workspace_dir=WORKSPACE_DIR, artifact_location=ARTIFACT_LOCATION, profile=PROFILE,
-            enable_inplace_jinja_support=False
+            enable_inplace_jinja_support=False, enable_failsafe_cluster_reuse_with_assets=False
         )
 
         env = Environment(loader=FileSystemLoader(COMPONENTS_PATH))