+++ tests/unit/api/launch/test_runners.py
@@ -1,5 +1,10 @@
-import pytest
+from unittest.mock import MagicMock
 
+import pytest
+from databricks_cli.sdk import JobsService
+from pytest_mock import MockerFixture
+
+from dbx.api.configure import ProjectConfigurationManager
 from dbx.api.launch.runners import RunSubmitLauncher
 from dbx.models.parameters.run_submit import RunSubmitV2d0ParamInfo, RunSubmitV2d1ParamInfo
 
@@ -44,3 +49,28 @@ def test_vd21_parameter_override_positive():
     parameters = RunSubmitV2d1ParamInfo(tasks=[{"task_key": "this", "python_wheel_task": {"parameters": ["b"]}}])
     RunSubmitLauncher.override_v2d1_parameters(spec, parameters)
     assert spec["tasks"][0]["python_wheel_task"]["parameters"] == ["b"]
+
+
+def test_run_submit_reuse(temp_project, mocker: MockerFixture):
+    ProjectConfigurationManager().enable_failsafe_cluster_reuse()
+    service_mock = mocker.patch.object(JobsService, "submit_run", MagicMock())
+    cluster_def = {"some_key": "some_value"}
+    mocker.patch(
+        "dbx.api.launch.runners.load_dbx_file",
+        MagicMock(
+            return_value={
+                "default": {
+                    "jobs": [
+                        {
+                            "name": "test",
+                            "job_clusters": [{"job_cluster_key": "some", "new_cluster": cluster_def}],
+                            "tasks": [{"task_key": "one", "job_cluster_key": "some"}],
+                        }
+                    ]
+                }
+            }
+        ),
+    )
+    launcher = RunSubmitLauncher(job="test", api_client=MagicMock(), deployment_run_id="aaa-bbb", environment="default")
+    launcher.launch()
+    service_mock.assert_called_once_with(tasks=[{"task_key": "one", "new_cluster": cluster_def}])